requirements.txt
Django
djangorestframework
django-filter
markdown
corsheaders
